+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - pytest tmpdir_factory support (bubenkoff)
 - depend on splinter 0.7.3, remove the previous status_code monkey patch (pelme)
+- add option `--splinter-wait-time` to specify splinter explicit wait timeout (pelme)
 
 1.4.6
 -----